Update and rename  word_count.py to word_count.py
